diag_cnt: Allow accepting counters in any format (dec/hex/octal)
Allow accepting counters in any format (dec/hex/octal), as this will
ease external tools to interface with mlx5ctl. Some tools (like Python
YAML parser) automatically translate hex values to decimal, and it is
also easier to call the command for a Python numeric list (which prints
the values in decimal) rather than on a list of strings which contains
hex values).

Signed-off-by: Leonid Bloch <lbloch@nvidia.com>